requirement.txt
joblib==1.1.0
numpy==1.21.4
opencv-python==4.5.4.60
pip==21.3.1
pyinstrument==4.1.1
scikit-learn==1.0.1
scipy==1.7.3
setuptools==57.0.0
threadpoolctl==3.0.0
wheel==0.36.2
pyserial==3.5